requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
amqp==5.2.0
appdirs==1.4.4
apsw==3.36.0.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
berserk==0.13.2
billiard==4.2.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
CacheControl==0.14.0
cachetools==5.3.3
certifi==2020.6.20
cffi==1.17.0
chardet==4.0.0
charset-normalizer==2.0.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.39.1
colorama==0.4.4
colorlog==6.8.2
command-not-found==0.3
cryptography==43.0.0
css-parser==1.0.7
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
Deprecated==1.2.14
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.1.0
docutils==0.17.1
feedparser==6.0.8
ffmpeg==1.4
ffmpeg-downloader==0.3.0
fonttools==4.29.1
frozenlist==1.2.0
fs==2.4.12
gast==0.5.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==1.1.2
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
howmanypeoplearearound==0.5.0
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
imagesize==1.3.0
importlib-metadata==4.6.4
ipython==7.31.1
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
kiwisolver==1.3.2
kombu==5.4.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mechanize==0.4.7
meson==0.61.2
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.8
multidict==5.1.0
ndjson==0.3.1
netifaces==0.11.0
numpy==2.1.0
oauthlib==3.2.0
olefile==0.46
packaging==24.1
parso==0.8.1
pexpect==4.8.0
pick==2.3.2
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.3.6
ply==3.11
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
proton-core==0.3.3
proton-keyring-linux==0.0.2
proton-keyring-linux-secretservice==0.0.2
proton-vpn-api-core==0.33.11
proton-vpn-connection==0.14.4
proton-vpn-gtk-app==4.4.5
proton-vpn-killswitch==0.4.0
proton-vpn-killswitch-network-manager==0.5.4
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.6.3
proton-vpn-network-manager-openvpn==0.1.1
ptyprocess==0.7.0
py7zr==0.11.3+dfsg
pyaml==24.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pychm==0.8.6
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
pydub==0.25.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gnupg==0.4.8
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
regex==2021.11.10
reportlab==3.6.8
repoze.lru==0.7
requests==2.31.0
requests-toolbelt==0.9.1
roman==3.3
Routes==2.5.1
rsa==4.9
scipy==1.8.0
SecretStorage==3.3.1
sentry-sdk==1.4.3
sgmllib3k==1.0.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.2
sqlparse==0.5.1
sympy==1.9
systemd-python==234
tabulate==0.8.9
texttable==1.6.4
tqdm==4.66.5
traitlets==5.1.1
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.5
uvicorn==0.30.4
vboxapi==1.0
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
wrapt==1.16.0
xdg==5
xkit==0.0.0
yarl==1.7.2
zeroconf==0.38.3
zipp==1.0.0
